+++ xwindow.h
@@ -25,7 +25,7 @@
 #include "globalconf.h"
 #include "draw.h"
 
-void xwindow_set_state(xcb_window_t, long);
+void xwindow_set_state(xcb_window_t, uint32_t);
 xcb_get_property_cookie_t xwindow_get_state_unchecked(xcb_window_t);
 uint32_t xwindow_get_state_reply(xcb_get_property_cookie_t);
 void xwindow_configure(xcb_window_t, area_t, int);
